backend/requirements_interview.txt
pyaudio==0.2.14
vosk==0.3.44
openai>=1.0.0
python-dotenv==1.0.0
